Move backquote parser to separate module
